move mapper files to a separate folder and update imports accordingly
