+++ src/modules/data/external-tool/index.ts
@@ -1,4 +1,3 @@
-export * from "./context-external-tool.mapper";
 export * from "./contextExternalToolApi.composable";
 export * from "./contextExternalToolConfigurationState.composable";
 export * from "./ContextExternalToolConfigurationStatus.composable";
@@ -7,6 +6,7 @@ export * from "./externalToolDisplayListState.composable";
 export * from "./externalToolDisplayState.composable";
 export * from "./ExternalToolLaunchState.composable";
 export * from "./externalToolReferenceApi.composable";
+export * from "./mapper";
 export * from "./preferred-external-tool.store";
 export * from "./school-external-tool-configurator.composable";
 export * from "./school-external-tools.composable";
