+++ src/modules/data/external-tool/mapper/external-tool.mapper.ts
@@ -1,5 +1,5 @@
+import { ToolParameter } from "../types/tool-parameter";
 import { ToolParamLocationMapping, ToolParamScopeMapping, ToolParamTypeMapping } from "./common-tool.mapper";
-import { ToolParameter } from "./tool-parameter";
 import { CustomParameterResponse } from "@api-server";
 
 export class ExternalToolMapper {
