+++ src/modules/data/external-tool/types/index.ts
@@ -1,5 +1,3 @@
-export * from "../types/external-tool.mapper";
-export * from "../types/school-external-tool.mapper";
 export * from "./context-external-tool";
 export * from "./external-tool-display-data";
 export * from "./school-external-tool";
